added telemetry stuff
